JSONStorage into master

+++ src/commands/ChatCommands.java
@@ -1,77 +1,91 @@
 package commands;
 
-import java.awt.Color;
-import java.io.File;
+import commands.moderation.Moderation;
+import commands.moderation.Permission;
+import db.DataManager;
+import main.MainBot;
+import main.Util;
+import sx.blah.discord.handle.impl.events.guild.channel.message.MessageReceivedEvent;
+import sx.blah.discord.util.EmbedBuilder;
+import sx.blah.discord.util.RequestBuffer;
+
+import java.awt.*;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import commands.moderation.Permission;
-import dataStore.DataStore;
-import main.MainBot;
-import main.Util;
-import org.slf4j.Logger;
-import sx.blah.discord.handle.impl.events.guild.channel.message.MessageReceivedEvent;
-import sx.blah.discord.util.EmbedBuilder;
-import sx.blah.discord.util.Image;
-import sx.blah.discord.util.RequestBuffer;
-
-import javax.imageio.ImageIO;
-import javax.imageio.ImageReader;
 
 public class ChatCommands {
 
   public static Map<String, ChatCommand> commandMap = new HashMap<>();
+  public static Map<String, ChatCommand> adminMap = new HashMap<>();
 
   public static void init() {
     ChatCommands.commandMap.put("rule", (event, args) -> {
-      //j.rule deny say everyone general
-      //j.rule allow say staff general
-      //j.rule meme deny everyone
-      //j.rule meme allow everyone botcommands
-      //j.rule satansbae deny jarza bot
-      //[deny or allow] [command name] (user default: everyone) (channel default: all) (role default: everyone)
-      DataStore.setPermission(Permission.commandParse(args, event.getGuild().getLongID()), event.getGuild().getLongID());
+      //j.rule say deny general jarza
+      // OR
+      //j.rule say deny jarza genearl
+      boolean cont = false;
+      for(Long a : DataManager.getModrole(event.getGuild().getLongID())){
+        if (event.getAuthor().getRolesForGuild(event.getGuild()).contains(event.getGuild().getRoleByID(a))){
+          cont = true;
+        }
+      }
+      if (!cont) {
+        Util.sendMessage(event, "**>Error: inssuficient permission**");
+        return;
+      }
+      if(DataManager.getPerms(event.getGuild().getLongID()).size() >= 25) {
+        Util.sendMessage(event, "Error: you can only have up to 25 permissions.");
+        return;
+      }
+      if(args.isEmpty())
+        return;
+      if(args.size() == 1)
+        return;
 
+
+      DataManager.setPermission(Permission.commandParse(args, event.getGuild().getLongID(), event), event.getGuild().getLongID());
     });
 
-    ChatCommands.commandMap.put("logoff", (event, args) -> {
+    ChatCommands.adminMap.put("logoff", (event, args) -> {
       if (event.getAuthor().getLongID() == Util.jarza) {
-        Util.sendMessage(event.getChannel(), "Turning off...");
-        MainBot.cli.logout();
+        Util.sendMessage(event, "Turning off...");
+        RequestBuffer.request(() -> MainBot.cli.logout());
+        DataManager.finish();
         System.exit(0);
       }
     });
 
-    ChatCommands.commandMap.put("godmode", (event, args) -> {
+    ChatCommands.adminMap.put("godmode", (event, args) -> {
       if (event.getAuthor().getLongID() == Util.jarza) {
         Util.gmode = !Util.gmode;
         if (Util.gmode) {
-          Util.sendMessage(event.getChannel(), "GodMode On - Only jaja can do commands now hehehe");
+          Util.sendMessage(event, "GodMode On - Only jaja can do commands now hehehe");
         } else {
-          Util.sendMessage(event.getChannel(), "GodMode Off - *just don't spam too much pls*");
+          Util.sendMessage(event, "GodMode Off - *just don't spam too much pls*");
         }
       }
     });
 
     ChatCommands.commandMap.put("invite", (event, args) -> {
-      Util.sendMessage(event.getChannel(), "Invite link for Jarzu botto: " + Util.link);
+      Util.sendMessage(event, "Invite link for Jarzu botto: " + Util.link);
     });
 
-    ChatCommands.commandMap.put("game", (MessageReceivedEvent event, List<String> args) -> {
+    ChatCommands.adminMap.put("game", (MessageReceivedEvent event, List<String> args) -> {
       if (event.getAuthor().getLongID() == Util.jarza) {
         if (args.size() != 0) {
           String text = "";
           for (Object a : args.toArray()) {
             text += " " + a;
           }
-          Util.sendMessage(event.getChannel(), "Changed game status to: " + text);
+          Util.sendMessage(event, "Changed game status to: " + text);
           MainBot.cli.changePlayingText(text);
         }
       }
     });
 
     commandMap.put("ping", (event, args) -> {
-      Util.sendMessage(event.getChannel(), "pong");
+      Util.sendMessage(event, "pong");
     });
 
     commandMap.put("info", (event, args) -> {
@@ -82,26 +96,54 @@ public class ChatCommands {
       builder.withColor(new Color(112, 137, 255));
       builder.withAuthorIcon(event.getAuthor().getAvatarURL());
       builder.withFooterText("Still under development! uwu");
-      builder.withThumbnail(MainBot.cli.getApplicationIconURL());
+      builder.withThumbnail(MainBot.cli.getOurUser().getAvatarURL());
 
-      RequestBuffer.request(() -> event.getChannel().sendMessage(builder.build()));
+      Util.sendMessage(event, builder.build());
     });
 
     commandMap.put("help", (event, args) -> {
       EmbedBuilder builder = new EmbedBuilder();
 
-      builder.appendField("Prefix: " + Util.prefix, String.format("Version %1$s", Util.version),
+      if(args.isEmpty()){
+        builder.appendField("Prefix: " + Util.prefix, String.format("Version %1$s", Util.version),
           true);
-      builder.withAuthorName("Jarza Bot Manual!");
-      builder.withAuthorIcon(event.getAuthor().getAvatarURL());
-      builder.withFooterText("Still under development! uwu");
-      builder.withColor(new Color(112, 137, 255));
-      for (int x = Util.cmdinfo.length; x > 0; x--) {
-        builder.appendField(Util.cmdinfo[Util.cmdinfo.length - x][0],
-            Util.cmdinfo[Util.cmdinfo.length - x][1], false);
-      }
+        builder.withAuthorName("Jarza Bot Manual!");
+        builder.withAuthorIcon(event.getAuthor().getAvatarURL());
+        builder.withFooterText("Still under development! uwu");
+        builder.withColor(new Color(112, 137, 255));
 
-      RequestBuffer.request(() -> event.getChannel().sendMessage(builder.build()));
+
+        for (String a : Util.catnames) {
+          String comms = "";
+          for(String c : Util.cats.get(a).keySet()) {
+            comms += c + ", ";
+          }
+          builder.appendField(a, comms, false);
+        }
+        Util.sendMessage(event, builder.build());
+      }else if(args.size() == 1){
+        builder.appendField("Prefix: " + Util.prefix, String.format("Version %1$s", Util.version),
+          true);
+        builder.withAuthorName("Jarza Bot Manual!");
+        builder.withAuthorIcon(event.getAuthor().getAvatarURL());
+        builder.withFooterText("Still under development! uwu");
+        builder.withColor(new Color(112, 137, 255));
+        String cmd = null, info = null;
+        for (String a : Util.catnames) {
+          for(String c : Util.cats.get(a).keySet()) {
+            if(c.equals(args.get(0).toLowerCase())){
+              cmd = c;
+              info = Util.cats.get(a).get(c);
+            }
+          }
+        }
+
+        if(cmd == null)
+          return;
+
+        builder.appendField(cmd, info, false);
+        Util.sendMessage(event, builder.build());
+      }
     });
 
     ChatCommands.commandMap.put("test", (event, args) -> {
@@ -116,11 +158,11 @@ public class ChatCommands {
       builder.withThumbnail(
           "https://cdn.discordapp.com/avatars/218787234910961665/34d21653c66436b94c2af78b30156e91.webp?size=64");
 
-      RequestBuffer.request(() -> event.getChannel().sendMessage(builder.build()));
+      Util.sendMessage(event, builder.build());
 //      MainBot.cli.changeAvatar(Image.forFile(new File("gfx/drawing.png")));
     });
 
-    commandMap.put("say", (event, args) -> {
+    commandMap.put("say", (MessageReceivedEvent event, List<String> args) -> {
       String text = "";
       {
         String t = "";
@@ -136,14 +178,28 @@ public class ChatCommands {
         }
       }
       System.out.println(event.getAuthor().getName() + " said: " + text);
-      event.getMessage().delete();
-      Util.sendMessage(event.getChannel(), text);
+      if(event.getGuild() != null)
+        event.getMessage().delete();
+
+      Util.sendMessage(event, text);
     });
 
     commandMap.put("satansbae", (event, args) -> {
       String song =
           "*Archangel*\n*Darkangel*\n*Bring us elevation*\n*Through hell and through heaven*\n***Until we reach ascention***";
-      RequestBuffer.request(() -> event.getChannel().sendMessage(song));
+      Util.sendMessage(event, song);
+    });
+
+    commandMap.put("disable", (event, args) -> {
+      if(args.isEmpty())
+        return;
+
+      Long g;
+      try{
+        g = Long.parseLong(args.get(0));
+      }catch(NumberFormatException e){return;}
+
+      DataManager.setUserNotifi(g, event.getAuthor().getLongID());
     });
   }
